LPS-63970 move configuration icons to the sharing web module
