--- modules/apps/web-experience-management/portlet-configuration/portlet-configuration-icon/portlet-configuration-icon-facebook/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Portlet Configuration Icon Facebook
-Bundle-SymbolicName: com.liferay.portlet.configuration.icon.facebook
-Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Site Management
--include: ../../../../../../marketplace/web-content-management/bnd.bnd
