--- modules/apps/web-experience-management/portlet-configuration/portlet-configuration-icon/portlet-configuration-icon-widget/build.gradle
@@ -1,11 +1,0 @@
-dependencies {
-	provided group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "1.0.0"
-	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "1.0.1"
-	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
-	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
-	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-}
-
-liferay {
-	deployDir = file("${liferayHome}/osgi/modules")
-}
