+++ modules/apps/web-experience-management/portlet-configuration/portlet-configuration-sharing-web/src/main/java/com/liferay/portlet/configuration/sharing/web/portlet/configuration/icon/WidgetPortletConfigurationIcon.java
@@ -12,25 +12,30 @@
  * details.
  */
 
-package com.liferay.portlet.configuration.icon.igoogle;
+package com.liferay.portlet.configuration.sharing.web.portlet.configuration.icon;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Portlet;
+import com.liferay.portal.kernel.portlet.PortletURLFactoryUtil;
 import com.liferay.portal.kernel.portlet.configuration.icon.BasePortletConfigurationIcon;
 import com.liferay.portal.kernel.portlet.configuration.icon.PortletConfigurationIcon;
 import com.liferay.portal.kernel.theme.PortletDisplay;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
+import com.liferay.portal.kernel.util.PortletKeys;
+import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.util.PropsValues;
 
 import javax.portlet.PortletPreferences;
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
+import javax.portlet.PortletURL;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -38,18 +43,13 @@ import org.osgi.service.component.annotations.Component;
  * @author Eudaldo Alonso
  */
 @Component(immediate = true, service = PortletConfigurationIcon.class)
-public class IGooglePortletConfigurationIcon
+public class WidgetPortletConfigurationIcon
 	extends BasePortletConfigurationIcon {
-
-	@Override
-	public String getCssClass() {
-		return "expose-as-widget";
-	}
 
 	@Override
 	public String getMessage(PortletRequest portletRequest) {
 		return LanguageUtil.get(
-			getResourceBundle(getLocale(portletRequest)), "add-to-igoogle");
+			getResourceBundle(getLocale(portletRequest)), "add-to-any-website");
 	}
 
 	@Override
@@ -64,8 +64,19 @@ public class IGooglePortletConfigurationIcon
 				(ThemeDisplay)portletRequest.getAttribute(
 					WebKeys.THEME_DISPLAY);
 
-			return "http://fusion.google.com/add?source=atgs&moduleurl=" +
-				PortalUtil.getGoogleGadgetURL(portlet, themeDisplay);
+			PortletURL basePortletURL = PortletURLFactoryUtil.create(
+				portletRequest, PortletKeys.PORTLET_SHARING,
+				themeDisplay.getPlid(), PortletRequest.RESOURCE_PHASE);
+
+			StringBundler sb = new StringBundler();
+
+			sb.append("javascript:Liferay.PortletSharing.showWidgetInfo('");
+			sb.append(PortalUtil.getWidgetURL(portlet, themeDisplay));
+			sb.append("', '");
+			sb.append(basePortletURL);
+			sb.append("');");
+
+			return sb.toString();
 		}
 		catch (PortalException pe) {
 			if (_log.isWarnEnabled()) {
@@ -78,7 +89,7 @@ public class IGooglePortletConfigurationIcon
 
 	@Override
 	public double getWeight() {
-		return 3.0;
+		return 5.0;
 	}
 
 	@Override
@@ -95,11 +106,11 @@ public class IGooglePortletConfigurationIcon
 
 		PortletPreferences portletSetup = portletDisplay.getPortletSetup();
 
-		boolean lfrIgoogleShowAddAppLink = GetterUtil.getBoolean(
-			portletSetup.getValue(
-				"lfrIgoogleShowAddAppLink", StringPool.BLANK));
+		boolean lfrWidgetShowAddAppLink = GetterUtil.getBoolean(
+			portletSetup.getValue("lfrWidgetShowAddAppLink", null),
+			PropsValues.THEME_PORTLET_SHARING_DEFAULT);
 
-		if (lfrIgoogleShowAddAppLink) {
+		if (lfrWidgetShowAddAppLink) {
 			return true;
 		}
 
@@ -107,6 +118,6 @@ public class IGooglePortletConfigurationIcon
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		IGooglePortletConfigurationIcon.class);
+		WidgetPortletConfigurationIcon.class);
 
 }
